# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LownieConfig {
  feeKes: number;
  currency: string;
  talkSasaReady: boolean;
  paystackReady: boolean;
}

export interface PhoneVerificationInput {
  /** @minLength 9 */
  phone: string;
}

export interface PhoneVerificationConfirmInput {
  /** @minLength 9 */
  phone: string;
  /**
     * @minLength 4
     * @maxLength 8
     */
  code: string;
  verificationId: string;
}

export interface VerificationStart {
  verificationId: string;
  message: string;
}

export interface VerificationResult {
  verified: boolean;
  phone: string;
  message: string;
}

export interface LownieApplicationInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 9 */
  phone: string;
  /** @minItems 1 */
  frustrations: string[];
  verificationId: string;
  consent: boolean;
}

export interface LownieApplication {
  applicationId: string;
  lownieId: string;
  name: string;
  phone: string;
  frustrations: string[];
  status: string;
  feeKes: number;
  paid: boolean;
  createdAt: string;
}

export interface PaymentInput {
  applicationId: string;
  email: string;
}

export interface PaymentInitialization {
  reference: string;
  authorizationUrl: string;
  accessCode: string;
}

export interface PaymentVerification {
  paid: boolean;
  reference: string;
  applicationId: string;
  message: string;
}