# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: lownie
    description: Lownie application workflow
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /lownie/config:
    get:
      operationId: getLownieConfig
      tags: [lownie]
      summary: Get the current Lownie fee and provider readiness
      responses:
        "200":
          description: Lownie configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LownieConfig"
  /lownie/verification/start:
    post:
      operationId: startPhoneVerification
      tags: [lownie]
      summary: Start phone verification with Talk Sasa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneVerificationInput"
      responses:
        "200":
          description: Verification started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationStart"
        "400":
          description: Invalid phone number
  /lownie/verification/confirm:
    post:
      operationId: confirmPhoneVerification
      tags: [lownie]
      summary: Confirm a phone verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneVerificationConfirmInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
        "400":
          description: Invalid verification code
  /lownie/applications:
    post:
      operationId: createLownieApplication
      tags: [lownie]
      summary: Create a Lownie application after phone verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LownieApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LownieApplication"
        "400":
          description: Invalid application
  /lownie/payments/initialize:
    post:
      operationId: initializeLowniePayment
      tags: [lownie]
      summary: Initialize the Paystack fee payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment initialization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitialization"
        "400":
          description: Invalid payment request
  /lownie/payments/verify/{reference}:
    get:
      operationId: verifyLowniePayment
      tags: [lownie]
      summary: Verify a Paystack payment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerification"
  /lownie/tickets/{applicationId}/download:
    get:
      operationId: downloadLownieTicket
      tags: [lownie]
      summary: Download a Lownie ID ticket
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Downloadable ticket document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Ticket not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LownieConfig:
      type: object
      properties:
        feeKes:
          type: number
        currency:
          type: string
        talkSasaReady:
          type: boolean
        paystackReady:
          type: boolean
      required: [feeKes, currency, talkSasaReady, paystackReady]
    PhoneVerificationInput:
      type: object
      properties:
        phone:
          type: string
          minLength: 9
      required: [phone]
    PhoneVerificationConfirmInput:
      type: object
      properties:
        phone:
          type: string
          minLength: 9
        code:
          type: string
          minLength: 4
          maxLength: 8
        verificationId:
          type: string
      required: [phone, code, verificationId]
    VerificationStart:
      type: object
      properties:
        verificationId:
          type: string
        message:
          type: string
      required: [verificationId, message]
    VerificationResult:
      type: object
      properties:
        verified:
          type: boolean
        phone:
          type: string
        message:
          type: string
      required: [verified, phone, message]
    LownieApplicationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 9
        frustrations:
          type: array
          items:
            type: string
          minItems: 1
        verificationId:
          type: string
        consent:
          type: boolean
      required: [name, phone, frustrations, verificationId, consent]
    LownieApplication:
      type: object
      properties:
        applicationId:
          type: string
        lownieId:
          type: string
        name:
          type: string
        phone:
          type: string
        frustrations:
          type: array
          items:
            type: string
        status:
          type: string
        feeKes:
          type: number
        paid:
          type: boolean
        createdAt:
          type: string
      required: [applicationId, lownieId, name, phone, frustrations, status, feeKes, paid, createdAt]
    PaymentInput:
      type: object
      properties:
        applicationId:
          type: string
        email:
          type: string
          format: email
      required: [applicationId, email]
    PaymentInitialization:
      type: object
      properties:
        reference:
          type: string
        authorizationUrl:
          type: string
        accessCode:
          type: string
      required: [reference, authorizationUrl, accessCode]
    PaymentVerification:
      type: object
      properties:
        paid:
          type: boolean
        reference:
          type: string
        applicationId:
          type: string
        message:
          type: string
      required: [paid, reference, applicationId, message]