# lib/api-zod/src/generated/types/lownieApplication.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LownieApplication {
  applicationId: string;
  lownieId: string;
  name: string;
  phone: string;
  frustrations: string[];
  status: string;
  feeKes: number;
  paid: boolean;
  createdAt: string;
}