# lib/api-zod/src/generated/types/lownieApplicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LownieApplicationInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 9 */
  phone: string;
  /** @minItems 1 */
  frustrations: string[];
  verificationId: string;
  consent: boolean;
}