# lib/api-zod/src/generated/types/lownieConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LownieConfig {
  feeKes: number;
  currency: string;
  talkSasaReady: boolean;
  paystackReady: boolean;
}